Update Sepolia Mention (#23)
* Updating Steps For Wallet Setup and Faucet

Updating the first paragraph of documentation to match the current branding of MultiversX and the number of tokens provided by the built-in faucet.

* Updating Steps for Wallet Setup and Faucet

Edited the first paragraph of the Try Our Flow section to match the new MultiversX branding + the number of tokens provided by the faucet.

* Updates To E-Mail Addresses + Integrations Links

Removed all @utrust.com e-mails for @xmoney.com addresses. Also cleaned the link for Opencart.

(Marketplace entries for WHMCS and Opencart were updated but their URLs are still the same.)

* Update Sepolia Mention

* Updates To E-Mail Addresses + Integrations Links

Removed all @utrust.com e-mails for @xmoney.com addresses. Also cleaned 
the link for Opencart.

(Marketplace entries for WHMCS and Opencart were updated but their URLs 
are still the same.)